+++ RW_UL/hoe/hoecRefactor/private/src/hoec_Camera_MVS_v1_private.cpp
@@ -548,7 +548,7 @@ namespace rw
 			return false;
 		}
 
-		bool Camera_MVS::getEncoderNumber(size_t& number)
+		bool Camera_MVS::getEncoderNumber(double& number)
 		{
 			return false;
 		}
